feat: support custom OpenAPI auth scheme (#51)
Currently, we inject an OIDC auth scheme into the OpenAPI spec. However,
users may want to run the STAC Auth Proxy to apply auth for tokens that
are generated elsewhere (e.g. tokens that can be validated with a JWKS,
but are not generated via the `/token` or `/authorization` endpoint). As
such, this PR enables the manual override of the auth scheme that we
inject into the OpenAPI doc, configurable via env vars.