fix: improve link processing (#95)
We expect that upstream STAC APIs will set the origin of URLs contained
in the `href` property of the `links` array to be properly pointed to
the STAC Auth Proxy URL based on the [`Forwarded`
header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Forwarded)
we provide. However, in certain configurations, upstream APIs don't
perform this change (e.g. the proxy headers are not properly forwarded
to the upstream STAC API via a reverse proxy). This PR updates how we
rework links to ensure that links that have an origin and path matching
the path of the upstream STAC API will be reworked to instead point to
the origin and potential path of the STAC Auth Proxy.